Tsing-li tso-hue.
