--- 臺灣言語工具/語音合成/閩南語音韻/變調/仔前變調.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-from 臺灣言語工具.語音合成.閩南語音韻.變調.規則變調 import 變調規則表
-
-
-仔前變調 = 變調規則表(
-    名='仔前變調',
-    喉入聲變調規則={'4': '1', '8': '7'},
-    入聲變調規則={'4': '8', '8': '4'},
-    變調規則={'1': '7', '2': '1', '3': '1', '5': '7', '7': '7'},
-)
