--- 臺灣言語工具/語音合成/閩南語音韻/變調/再變調.py
@@ -1,9 +1,0 @@
-# -*- coding: utf-8 -*-
-from 臺灣言語工具.語音合成.閩南語音韻.變調.規則變調 import 規則變調
-
-
-class 再變調:
-
-    @classmethod
-    def 變調(cls, 音):
-        return 規則變調.變調(規則變調.變調(音))
