--- 臺灣言語工具/語音合成/閩南語音韻/變調/無調符號.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-
-
-class 無調符號:
-
-    @classmethod
-    def 變調(cls, 音):
-        return 音
