--- 臺灣言語工具/語音合成/閩南語音韻/變調/輕聲.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-from 臺灣言語工具.語音合成.閩南語音韻.變調.規則變調 import 變調規則表
-
-
-輕聲 = 變調規則表(
-    名='輕聲',
-    喉入聲變調規則={'4': '3', '8': '3', '0': '3'},
-    入聲變調規則={'4': '10', '8': '10', '0': '10'},
-    變調規則={'1': '3', '2': '3', '3': '3', '5': '3', '7': '3', '0': '3'},
-)
