+++ 臺灣言語工具/語音合成/閩南語音韻/變調/類型.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from 臺灣言語工具.解析整理.解析錯誤 import 解析錯誤
 
 
@@ -28,11 +27,3 @@ class 變調規則表:
             else:
                 raise 解析錯誤('非入調錯誤！！{0}'.format((聲, 韻, 調)))
         return (聲, 韻, 調)
-
-
-規則變調 = 變調規則表(
-    名='規則變調',
-    喉入聲變調規則={'4': '2', '8': '3'},
-    入聲變調規則={'4': '8', '8': '10'},
-    變調規則={'1': '7', '2': '1', '3': '2', '5': '7', '7': '3', '9': '9'},
-)
