+++ 臺灣言語工具/語音合成/閩南語音韻/變調判斷.py
@@ -2,14 +2,14 @@
 from 臺灣言語工具.基本物件.公用變數 import 本調符號
 from 臺灣言語工具.基本物件.句 import 句
 from 臺灣言語工具.基本物件.章 import 章
-from 臺灣言語工具.語音合成.閩南語音韻.變調.規則變調 import 規則變調
-from 臺灣言語工具.語音合成.閩南語音韻.變調.維持本調 import 維持本調
-from 臺灣言語工具.語音合成.閩南語音韻.變調.無調符號 import 無調符號
-from 臺灣言語工具.語音合成.閩南語音韻.變調.輕聲 import 輕聲
-from 臺灣言語工具.語音合成.閩南語音韻.變調.三連音變調 import 三連音變調
-from 臺灣言語工具.語音合成.閩南語音韻.變調.仔前變調 import 仔前變調
-from 臺灣言語工具.語音合成.閩南語音韻.變調.隨前變調 import 隨前變調
-from 臺灣言語工具.語音合成.閩南語音韻.變調.再變調 import 再變調
+from 臺灣言語工具.語音合成.閩南語音韻.變調 import 規則變調
+from 臺灣言語工具.語音合成.閩南語音韻.變調 import 維持本調
+from 臺灣言語工具.語音合成.閩南語音韻.變調 import 無調符號
+from 臺灣言語工具.語音合成.閩南語音韻.變調 import 輕聲
+from 臺灣言語工具.語音合成.閩南語音韻.變調 import 三連音變調
+from 臺灣言語工具.語音合成.閩南語音韻.變調 import 仔前變調
+from 臺灣言語工具.語音合成.閩南語音韻.變調 import 隨前變調
+from 臺灣言語工具.語音合成.閩南語音韻.變調 import 再變調
 
 
 class 變調判斷:
