+++ 試驗/語音合成/閩南語音韻/Test變調判斷本調單元試驗.py
@@ -3,7 +3,7 @@ from unittest.case import TestCase
 from 臺灣言語工具.解析整理.拆文分析器 import 拆文分析器
 from 臺灣言語工具.音標系統.閩南語.臺灣閩南語羅馬字拼音 import 臺灣閩南語羅馬字拼音
 from 臺灣言語工具.語音合成.閩南語音韻.變調判斷 import 變調判斷
-from 臺灣言語工具.語音合成.閩南語音韻.變調.維持本調 import 維持本調
+from 臺灣言語工具.語音合成.閩南語音韻.變調 import 維持本調
 
 
 class 變調判斷本調單元試驗(TestCase):
